fixup! Use regexps from git_helper
